pneumo-ai-pipeline/requirements-eval.txt
filelock==3.18.0
fsspec==2025.7.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
pillow==11.3.0
setuptools==80.9.0
sympy==1.14.0
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
typing_extensions==4.14.1
matplotlib>=3.7.1
seaborn>=0.12.0
scikit-learn>=1.3.0
boto3>=1.28.0
PyYAML>=6.0.1
python-dotenv>=1.0.0
tqdm>=4.65.0